moved lexer out into it's own header and cpp file
